Increase hidden layer size to 1280 and add 4 input buckets (#75)
Bench: 2250744

+++ src/utils/list.h
@@ -57,7 +57,7 @@ class List {
 
  private:
   std::array<T, length> container_;
-  std::uint8_t count_ = 0;
+  std::int32_t count_ = 0;
 };
 
 #endif  // INTEGRAL_LIST_H
